fix(test): cast sqlite3_column_int64 to int64_t in session store test
CPPUNIT_ASSERT_EQUAL is a single-type template (assertEquals<T>). On LP64
platforms int64_t is `long` while sqlite3_int64 is `long long`, so comparing
`createdAt` (int64_t) against sqlite3_column_int64()'s return deduced
conflicting types and failed to compile, breaking `make check`. Cast the
column value to int64_t to match. Pre-existing since the sqlite3 session-store
feature; independent of the dependency-version work.

+++ test/Sqlite3SessionStoreTest.cc
@@ -321,7 +321,7 @@ void Sqlite3SessionStoreTest::testUpsertTaskInsertsThenUpdates()
         reinterpret_cast<const char*>(sqlite3_column_text(stmt, 2));
     CPPUNIT_ASSERT_EQUAL(std::string("paused"),
                          std::string(state ? state : ""));
-    CPPUNIT_ASSERT_EQUAL(createdAt, sqlite3_column_int64(stmt, 3));
+    CPPUNIT_ASSERT_EQUAL(createdAt, static_cast<int64_t>(sqlite3_column_int64(stmt, 3)));
     sqlite3_finalize(stmt);
   }
 }
